Nest index removal in if statement.

+++ db/migrate/20250220162254_fix_instruments_name_index.rb
@@ -2,6 +2,8 @@
 
 class FixInstrumentsNameIndex < ActiveRecord::Migration[6.1]
   def change
-    remove_index :instruments, name: 'index_instruments_on_lower_name_and_song'
+    if index.exists?(:instruments, name: 'index_instruments_on_lower_name_and_song')
+      remove_index :instruments, name: 'index_instruments_on_lower_name_and_song'
+    end
   end
 end
